HARD U-turn on the div idea.
I subsequently learned a lot about DOM manipulation and what NOT to do.
So time spent learing is not wasted!

Changed the DIVs to Button elements.

Created HTML Data attributes for all buttons to reference in JS.

Started creating the Calculator class with some methods, currently trying
to get the numbers to display on the display when clicked.

Succesfully managed to get an '[OBJECT HMTL DIV ELEMENT] to go there,

so its 50% of the way there!!